+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "result_management",
+  "version": "1.0.0",
+  "description": "a result management system",
+  "main": "server.js",
+  "scripts": {
+    "start": "nodemon backend/index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/snehil-01/Result_Management.git"
+  },
+  "author": "snehil vispute and tanmay raghuwanshi",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/snehil-01/Result_Management/issues"
+  },
+  "homepage": "https://github.com/snehil-01/Result_Management#readme",
+  "dependencies": {
+    "dotenv": "^16.0.2",
+    "express": "^4.18.2",
+    "mongoose": "^6.6.7",
+    "nodemon": "^2.0.20",
+    "path": "^0.12.7"
+  }
+}
